[NEW] Padronizacao na disposicao dos campos de endereco da tela de cadastro de bancos.
