+++ br_base/models/__init__.py
@@ -7,3 +7,4 @@ from . import br_base
 from . import res_company
 from . import res_country
 from . import res_partner
+from . import res_bank
