+++ br_base/views/res_company_view.xml
@@ -14,7 +14,7 @@
             <field name="street2" position="replace"/>
             <field name="zip" position="replace"/>
             <field name="street" position="replace">
-                <field name="zip" placeholder="Cep" class="o_address_zip" />
+                <field name="zip" placeholder="CEP" class="o_address_zip" />
                 <field name="street" placeholder="Logradouro" class="o_address_street" />
                 <field name="number" placeholder="Número" class="o_address_city"/>
                 <field name="street2" placeholder="Complemento" class="o_address_street" />
